Switched to correct gender of "Token"
The neutral gender "das Token" is used in science to describe an array of bits or chars, where as the masculin "der Token" is commenly used for keys in communications and authentication in the network.

I guess the a CSRF-Token is more a key for authentication then a bit array.