[Form] Fix merging params & files when "multiple" is enabled
If a form field is configured with the "multiple" option, the request
handler merges param values with uploaded files. However, it does not
merge correctly if "multiple" is enabled.

Example:

    <input type="file" name="param[]" />
    <input type="hidden" name="param[]" value="data1" />
    <input type="hidden" name="param[]" value="data2" />

With the example, the file field will incorrectly replace the first
hidden field, when they should be merged. This commit fixes the problem.